+++ eslint.config.js
@@ -4,25 +4,7 @@ import tsPlugin from "typescript-eslint";
 import prettierOverrides from "eslint-config-prettier";
 import prettierRules from "eslint-plugin-prettier/recommended";
 import unicornPlugin from "eslint-plugin-unicorn";
-
-import { readFile } from "node:fs/promises";
-import { reject, startsWith, partition, path, flip } from "ramda";
-
-// @todo consider "import with" starting Node 18.20 and 20.10
-const manifest = JSON.parse(await readFile("./package.json", "utf-8"));
-
-const unlistedPeers = ["eslint", "prettier"];
-const excludeTypes = reject(startsWith("@types/"));
-const lookup = flip(path)(manifest);
-const allPeers = excludeTypes(Object.keys(manifest.peerDependencies));
-const isOptional = (name) => lookup(["peerDependenciesMeta", name, "optional"]);
-const [optPeers, reqPeers] = partition(isOptional, allPeers);
-const production = Object.keys(manifest.dependencies);
-const allowed = production.concat(reqPeers);
-const typeOnly = optPeers.concat(unlistedPeers);
-
-console.debug("Allowed imports", allowed);
-console.debug("Type only imports", typeOnly);
+import { allowed, typeOnly } from "./tools/allowed-deps.js";
 
 export default [
   {
